refactor: domain 패키지 분리
- 도메인 객체들을 attendance/domain 패키지에 보관한다.
