+++ src/main/java/attendance/Application.java
@@ -1,0 +1,4 @@
+package attendance;
+
+public class Application {
+}
