+++ src/main/java/attendance/domain/AttendanceHistory.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.domain;
 
 import java.time.LocalDate;
 import java.util.HashSet;
