+++ src/main/java/attendance/domain/Crew.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.domain;
 
 import java.util.Objects;
