+++ src/test/java/attendance/AttendanceHistoryTest.java
@@ -3,6 +3,8 @@ package attendance;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import attendance.domain.Attendance;
+import attendance.domain.AttendanceHistory;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Optional;
