+++ src/test/java/attendance/AttendancePolicyTest.java
@@ -2,6 +2,7 @@ package attendance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import attendance.domain.AttendancePolicy;
 import java.time.DayOfWeek;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
