+++ src/test/java/attendance/CampusManagerTest.java
@@ -2,6 +2,7 @@ package attendance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import attendance.domain.CampusManager;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
