+++ src/test/java/attendance/CrewTest.java
@@ -2,6 +2,7 @@ package attendance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import attendance.domain.Crew;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
